+++ src/SMAPI.Installer/Framework/InstallerContext.cs
@@ -54,5 +54,12 @@ namespace StardewModdingAPI.Installer.Framework
         {
             return this.GameScanner.LooksLikeGameFolder(dir);
         }
+
+        /// <summary>Get whether a folder seems to contain Stardew Valley 1.5.4 or earlier.</summary>
+        /// <param name="dir">The folder to check.</param>
+        public bool LooksLikeStardewValley154(DirectoryInfo dir)
+        {
+            return this.GameScanner.LooksLikeStardewValley154(dir);
+        }
     }
 }